backend/requirements.txt
# HuggingFace Inference API
huggingface-hub>=0.16.0
huggingface-inference>=0.28.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Image Processing
Pillow>=9.0.0

# LLM API
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0